+++ apps/backend/src/Application/Families/Specifications/FamilyAccessSpecification.cs
@@ -1,5 +1,4 @@
 using Ardalis.Specification;
-using backend.Application.Common.Interfaces;
 using backend.Domain.Entities;
 
 namespace backend.Application.Families.Specifications;
