+++ apps/backend/src/Application/FamilyDicts/Commands/DeleteFamilyDict/DeleteFamilyDictCommand.cs
@@ -1,5 +1,4 @@
 using backend.Application.Common.Models; // Added
-using MediatR; // Added
 
 namespace backend.Application.FamilyDicts.Commands.DeleteFamilyDict;
